Move IECoreScenePreview type ids to separate header
